Internally prefixes are now parallel
The semantics for prefixes was implicit, there was
no distinction between `Act ps [Act qs procs]`
and `Act (ps ++ qs) procs`.

Now `Act` means sequencing (and will be renamed Dot for instance).
The prefix (a list of actions) is now explicitly in parallel.
Nesting is thus necessary to express multiple sequencing:
  `Act [a0] [Act [a1] [Act [a2] []]]`

The concrete syntax has not changed yet and thus those not have
any `.` to express sequencing nor `|` to express parallelism.
The translation for `a0 a1 a2` is nesting as the example above.
Finally this means that the implicit parallelism of prefixes is
temporarily disabled.