More flexible telescopes (close Issue#12)
At first one would expect the following production:

  VarsDec ::= "(" Name [Name] ":" Term ")"

However this falls in the restrictions of LALR and we can no
longer parse terms such as `(a b)`. Instead of switching
to GLR, one can tweak the grammar to accept any RawApp:

  VarsDec ::= "(" ATerm [ATerm] ":" Term ")"

The extra terms are rejected after in Ling.Reify.

Additionally:
* The syntax for `case` has now a slightly higher priority.
* The syntax for `slice` and `fwd` now take a list of `ChanDec`
  instead of `Name`. However these session annotations are not
  yet handled but rejected.
* Updated tests