Major refactoring of the Proc type
The main motivation was to make NewSlice a proper constructor
of Proc. Having NewSlice as an action was wrong and blocking.
Moreover the old Proc type was not supporting arbitrary Dot
between processes. The Sequential module was thus abusively using
the parallel syntax.

The new Proc type is as follows:

```
data Proc
  = Act Act
  | Procs (Prll Proc)
  | Dot Proc Proc
  | NewSlice [Channel] RFactor Name Proc
```

The `Prll` type is a newtype around `[]` suggesting that parallel
composition is applied.

A prism `_PrefDotProc` is exposed to partly recover the old layout
of processes which was convenient at times.

The new representation for processes is supported by a `Dottable`
type class which helps combining actions, prefixes and processes
in a sequential fashion.

This commit also improves the Sequential module by teaching it how
to unfold definitions and use different locations when stepping
through a continued session.